requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Approximate nearest neighbor search
annoy>=1.17.3

# Web demo
streamlit>=1.28.0

# Database (Firebase Firestore)
firebase-admin>=6.0.0

# NLP and embeddings
spacy>=3.7.0
sentence-transformers>=2.2.0

# BM25 for hybrid retrieval (lexical search)
rank-bm25>=0.2.2

# PDF and DOCX parsing (required for file upload feature)
PyPDF2>=3.0.0
python-docx>=1.0.0

# Analytics and Visualization
wordcloud>=1.8.1
matplotlib>=3.3.0
seaborn>=0.11.0

# PyTorch (required for sentence-transformers and fine-tuning)
torch>=2.0.0

# API Framework (Phase 3: REST API Layer)
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Authentication & Security (Phase 3)
PyJWT==2.8.0
python-multipart==0.0.6

# Background Tasks (Phase 3 - Optional but recommended for production)
celery==5.3.6
redis==5.0.1

# Report Generation (Phase 4)
openpyxl==3.1.2
reportlab==4.0.7

# Development
pytest>=7.4.0
